Return 2024 Grain-rain

+++ inc/c/uctype.h
@@ -51,9 +51,9 @@
 extern const unsigned char _tab_tolower[];
 extern const unsigned char _tab_toupper[];
 
-#define ascii_tolower(c) (_tab_tolower[c]) //((c)-'A'<26?(c)|0x20:c)
+#define ascii_tolower(c) (_tab_tolower[(byte)c]) //((c)-'A'<26?(c)|0x20:c)
 
-#define ascii_toupper(c) (_tab_toupper[c]) //((c)-'a'<26?(c)&~0x20:c)
+#define ascii_toupper(c) (_tab_toupper[(byte)c]) //((c)-'a'<26?(c)&~0x20:c)
 
 #define ascii_tohexad(c) ((c)>='a'?(c)-'a'+10:(c)>='A'?(c)-'A'+10 :(c)-'0')
